Fix Issue #4 and #7: Update Class Names in 'styles.css' and Remove ':disabled' Override in Favour of Plugin-Specific Class
To prevent conflicts with Obsidian and other plugins, change the names of classes to be more unique by prefixing them with the name of the plugin.
Another conflict includes the use of the ':disabled' pseudo class. Change it to use the 'setDisabled' method in Obsidian API. This makes it much easier to control the style of disabled elements.
Make the style of settings follow the default style of Obsidian.
Refactor 'settings.ts' so that all settings have a style for eaiser control of styling in the future.
Replace buttons with icon buttons with 'cross' symbol for simpler design.